MNT bump array-api-extra to v0.11.2
