+++ maint_tools/vendor_array_api_extra.sh
@@ -6,7 +6,7 @@ set -o nounset
 set -o errexit
 
 URL="https://github.com/data-apis/array-api-extra.git"
-VERSION="v0.11.1"
+VERSION="v0.11.2"
 
 ROOT_DIR=sklearn/externals/array_api_extra
