+++ sklearn/externals/array_api_extra/_agnostic/__init__.py
@@ -1,0 +1,27 @@
+"""Array-agnostic function implementations."""
+
+from . import (
+    _creation,
+    _elementwise,
+    _indexing,
+    _inspection,
+    _linalg,
+    _manipulation,
+    _searching,
+    _set,
+    _sorting,
+    _statistical,
+)
+
+__all__ = [
+    "_creation",
+    "_elementwise",
+    "_indexing",
+    "_inspection",
+    "_linalg",
+    "_manipulation",
+    "_searching",
+    "_set",
+    "_sorting",
+    "_statistical",
+]
