+++ sklearn/externals/array_api_extra/_lib/_compat.py
@@ -1,10 +1,10 @@
-"""Acquire helpers from array-api-compat."""
+"""Helpers from array-api-compat."""
 # Allow packages that vendor both `array-api-extra` and
 # `array-api-compat` to override the import location
 
 # pylint: disable=duplicate-code
 try:
-    from ...._array_api_compat_vendor import (
+    from ..._array_api_compat_vendor import (
         array_namespace,
         device,
         is_array_api_obj,
