--- sklearn/externals/array_api_extra/_lib/_utils/__init__.py
@@ -1,1 +1,0 @@
-"""Modules housing private utility functions."""
