--- sklearn/externals/array_api_extra/_lib/_utils/_typing.py
@@ -1,13 +1,0 @@
-# numpydoc ignore=GL08
-# pylint: disable=missing-module-docstring,duplicate-code
-
-from types import ModuleType
-
-Array = object
-ArrayNamespace = ModuleType
-DType = object
-Device = object
-GetIndex = object
-SetIndex = object
-
-__all__ = ["Array", "ArrayNamespace", "DType", "Device", "GetIndex", "SetIndex"]
