+++ sklearn/externals/array_api_extra/testing/__init__.py
@@ -1,0 +1,19 @@
+"""Public testing utilities."""
+
+from ._testing import (
+    assert_close,
+    assert_close_nulp,
+    assert_equal,
+    assert_less,
+    lazy_xp_function,
+    patch_lazy_xp_functions,
+)
+
+__all__ = [
+    "assert_close",
+    "assert_close_nulp",
+    "assert_equal",
+    "assert_less",
+    "lazy_xp_function",
+    "patch_lazy_xp_functions",
+]
